+++ src/TRegx/CleanRegex/Match/AbstractMatchPattern.php
@@ -44,6 +44,13 @@ abstract class AbstractMatchPattern implements MatchPatternInterface, PatternLim
         $this->base = $base;
     }
 
+    abstract public function test(): bool;
+
+    public function fails(): bool
+    {
+        return !$this->test();
+    }
+
     public function all(): array
     {
         return $this->base->matchAll()->getTexts();