+++ test/Integration/TRegx/CleanRegex/Match/FilteredMatchPattern/matches/FilteredMatchPatternTest.php
@@ -23,9 +23,11 @@ class FilteredMatchPatternTest extends TestCase
 
         // when
         $matches = $matchPattern->test();
+        $fails = $matchPattern->fails();
 
         // then
         $this->assertTrue($matches);
+        $this->assertFalse($fails);
     }
 
     /**
@@ -38,9 +40,11 @@ class FilteredMatchPatternTest extends TestCase
 
         // when
         $matches = $matchPattern->test();
+        $fails = $matchPattern->fails();
 
         // then
         $this->assertTrue($matches);
+        $this->assertFalse($fails);
     }
 
     /**
@@ -53,9 +57,11 @@ class FilteredMatchPatternTest extends TestCase
 
         // when
         $matches = $matchPattern->test();
+        $fails = $matchPattern->fails();
 
         // then
         $this->assertFalse($matches);
+        $this->assertTrue($fails);
     }
 
     /**
@@ -68,9 +74,11 @@ class FilteredMatchPatternTest extends TestCase
 
         // when
         $matches = $matchPattern->test();
+        $fails = $matchPattern->fails();
 
         // then
         $this->assertFalse($matches);
+        $this->assertTrue($fails);
     }
 
     private function standardMatchPattern(): FilteredMatchPattern